refactor(orchestrator): extract maintenance scheduling into MaintenanceScheduler (#1526 PR1)
First extraction under the #1526 extract-when-touched policy. Moves the
"when" of memory maintenance off the orchestrator god file into a new
orchestration/maintenance.ts module with an explicit, constructor-injected
interface and its own colocated suite.

Moved into MaintenanceScheduler (behavior-preserving — no logic changes):
- Cron auto-registration (day-summary, nightly-governance, procedural-mining,
  contradiction-scan, pattern-reinforcement, graph-edge-decay) — one
  autoRegisterCrons() entry replacing the orchestrator's sequential try/catch
  block, each sub-registration still individually gated + non-fatal.
- Debounced + singleflight QMD index maintenance (requestQmdMaintenance,
  requestQmdMaintenanceForTool, runQmdMaintenance) including the namespace
  planner path that unions configured + cataloged dynamic namespaces.
- Consolidation scheduling cadence state (nonZeroExtractionsSinceConsolidation,
  lastConsolidationRunAtMs, consolidationInFlight) — consolidation execution
  stays on the orchestrator (LLM-heavy) and is invoked through a callback.

Orchestrator keeps thin delegating methods (public API unchanged):
requestQmdMaintenance, requestQmdMaintenanceForTool, maybeScheduleConsolidation
now delegate to the scheduler. dispose() tears down the scheduler's timers.
waitForConsolidationIdle polls the scheduler via isConsolidationInFlight().

Cleanup + test relocation:
- Removed a dead leftover qmdMaintenanceSkipReasonForError copy from
  orchestrator.ts (the live copy moved with the scheduler; the leftover
  referenced a type import the extraction had already dropped → compile error).
- The 10 runQmdMaintenance unit tests moved from orchestrator-flush.test.ts
  into the new colocated orchestration/maintenance.test.ts, now constructing
  MaintenanceScheduler directly and driving cadence state through the public
  requestQmdMaintenanceForTool path (no private-field poking).
- The orchestrator-characterization.test.ts union pin now exercises the
  scheduler directly (the behavior moved with it).
- day-summary-cron-routing.test.ts calls the scheduler's public
  autoRegisterDaySummaryCron via the orchestrator's typed maintenanceScheduler
  field; orchestrator-nightly-governance-cron.test.ts retargets its per-cron
  stubs to the scheduler.
- Added @remnic/core/orchestration/maintenance export path + src/ shim so the
  scheduler is importable from root tests like every other module.

Gates: tsc clean (core); check-test-types matches baseline (0 new errors);
10/10 maintenance + 10/10 cron + 35/35 flush + 246/246 entity-hardening pass;
ratchet orchestrator.ts 19946 -> 19556 (locked via --update).

Chokepoints used: serialize-mutations/withHeldFileLock (unchanged — the
scheduler consumes the planner's locks); no new config flags. Not applicable:
no new MCP/HTTP/CLI ops.
